refactor: PostList 컴포넌트에 사용된 util 함수 분리
